Introduce revised useAdministrationsQuery composable
